requirements.txt
django
djangorestframework
django-filter
postgres
psycopg2
django-environ
django-rest-knox
django-crispy-forms